+++ docs/api.rst
@@ -49,7 +49,7 @@ developers, not a gospel.
     api/pulp_smash.tests.rpm.api_v2
     api/pulp_smash.tests.rpm.api_v2.test_broker
     api/pulp_smash.tests.rpm.api_v2.test_comps_groups
-    api/pulp_smash.tests.rpm.api_v2.test_download
+    api/pulp_smash.tests.rpm.api_v2.test_download_policies
     api/pulp_smash.tests.rpm.api_v2.test_duplicate_uploads
     api/pulp_smash.tests.rpm.api_v2.test_export
     api/pulp_smash.tests.rpm.api_v2.test_iso_crud
