--- docs/api/pulp_smash.tests.rpm.api_v2.test_download.rst
@@ -1,7 +1,0 @@
-`pulp_smash.tests.rpm.api_v2.test_download`
-===========================================
-
-Location: :doc:`/index` → :doc:`/api` →
-:doc:`/api/pulp_smash.tests.rpm.api_v2.test_download`
-
-.. automodule:: pulp_smash.tests.rpm.api_v2.test_download
