+++ docs/api/pulp_smash.tests.rpm.api_v2.test_download_policies.rst
@@ -1,0 +1,7 @@
+`pulp_smash.tests.rpm.api_v2.test_download_policies`
+====================================================
+
+Location: :doc:`/index` → :doc:`/api` →
+:doc:`/api/pulp_smash.tests.rpm.api_v2.test_download_policies`
+
+.. automodule:: pulp_smash.tests.rpm.api_v2.test_download_policies
