+++ pulp_smash/tests/rpm/api_v2/test_sync_publish.py
@@ -31,11 +31,9 @@ Assertions not explored in this module include:
 """
 from __future__ import unicode_literals
 
-import hashlib
 from itertools import product
 
 import unittest2
-from packaging.version import Version
 
 from pulp_smash import api, selectors, utils
 from pulp_smash.compat import urljoin
@@ -45,7 +43,6 @@ from pulp_smash.constants import (
     REPOSITORY_PATH,
     RPM,
     RPM_FEED_URL,
-    RPM_SHA256_CHECKSUM,
 )
 from pulp_smash.tests.rpm.api_v2.utils import gen_distributor, gen_repo
 from pulp_smash.tests.rpm.utils import set_up_module as setUpModule  # noqa pylint:disable=unused-import
@@ -395,190 +392,6 @@ class PublishTestCase(utils.BaseAPITestCase):
         for i, module in enumerate(self.rpms[1:]):
             with self.subTest(i=i):
                 self.assertEqual(self.rpms[0], module)
-
-
-class SyncOnDemandTestCase(utils.BaseAPITestCase):
-    """Assert the RPM plugin supports on-demand syncing of yum repositories.
-
-    Beware that this test case will fail if Pulp's Squid server is not
-    configured to return an appropriate hostname or IP when performing
-    redirection.
-    """
-
-    @classmethod
-    def setUpClass(cls):
-        """Create an RPM repository with a valid feed and sync it.
-
-        Do the following:
-
-        1. Reset Pulp, including the Squid cache.
-        2. Create a repository. Sync and publish it using the 'on_demand'
-           download policy.
-        3. Download an RPM from the published repository.
-        4. Download the same RPM to ensure it is served by the cache.
-        """
-        super(SyncOnDemandTestCase, cls).setUpClass()
-        if cls.cfg.version < Version('2.8'):
-            raise unittest2.SkipTest('This test requires Pulp 2.8 or greater.')
-
-        # Ensure `locally_stored_units` is 0 before we start.
-        utils.reset_squid(cls.cfg)
-        utils.reset_pulp(cls.cfg)
-
-        # Create a repository
-        client = api.Client(cls.cfg, api.json_handler)
-        body = gen_repo()
-        body['importer_config'] = {
-            'download_policy': 'on_demand',
-            'feed': RPM_FEED_URL,
-        }
-        distributor = gen_distributor()
-        distributor['auto_publish'] = True
-        distributor['distributor_config']['relative_url'] = body['id']
-        body['distributors'] = [distributor]
-
-        repo = client.post(REPOSITORY_PATH, body)
-        cls.resources.add(repo['_href'])
-
-        # Sync and read the repository
-        sync_path = urljoin(repo['_href'], 'actions/sync/')
-        client.post(sync_path, {'override_config': {}})
-        cls.repo = client.get(repo['_href'], params={'details': True})
-
-        # Download the same RPM twice.
-        client.response_handler = api.safe_handler
-        path = urljoin('/pulp/repos/', repo['id'] + '/')
-        path = urljoin(path, RPM)
-        cls.rpm = client.get(path)
-        cls.same_rpm = client.get(path)
-
-    def test_local_units(self):
-        """Assert no content units were downloaded besides metadata."""
-        metadata_unit_count = sum([
-            count for name, count in self.repo['content_unit_counts'].items()
-            if name not in ('rpm', 'drpm', 'srpm')
-        ])
-        self.assertEqual(
-            self.repo['locally_stored_units'],
-            metadata_unit_count
-        )
-
-    def test_repository_units(self):
-        """Assert there is at least one content unit in the repository."""
-        total_units = sum(self.repo['content_unit_counts'].values())
-        self.assertEqual(self.repo['total_repository_units'], total_units)
-
-    def test_request_history(self):
-        """Assert the initial request received a 302 Redirect."""
-        self.assertTrue(self.rpm.history[0].is_redirect)
-
-    def test_rpm_checksum(self):
-        """Assert the checksum of the downloaded RPM matches the metadata."""
-        checksum = hashlib.sha256(self.rpm.content).hexdigest()
-        self.assertEqual(RPM_SHA256_CHECKSUM, checksum)
-
-    def test_rpm_cache_lookup_header(self):
-        """Assert the first request resulted in a cache miss from Squid."""
-        self.assertIn('MISS', self.rpm.headers['X-Cache-Lookup'])
-
-    def test_rpm_cache_control_header(self):
-        """Assert the request has the Cache-Control header set."""
-        self.assertEqual(
-            {'s-maxage=86400', 'public', 'max-age=86400'},
-            set(self.rpm.headers['Cache-Control'].split(', '))
-        )
-
-    def test_same_rpm_checksum(self):
-        """Assert the checksum of the second RPM matches the metadata."""
-        checksum = hashlib.sha256(self.same_rpm.content).hexdigest()
-        self.assertEqual(RPM_SHA256_CHECKSUM, checksum)
-
-    def test_same_rpm_cache_header(self):
-        """Assert the second request resulted in a cache hit from Squid."""
-        self.assertIn('HIT', self.same_rpm.headers['X-Cache-Lookup'])
-
-
-class SyncBackgroundTestCase(utils.BaseAPITestCase):
-    """Assert Yum supports background syncing."""
-
-    @classmethod
-    def setUpClass(cls):
-        """Create an RPM repository with a valid feed and sync it.
-
-        Do the following:
-
-        0. Reset Pulp, including the Squid cache.
-        1. Sync/publish a repository using the 'background' download policy.
-        2. Assert all the content is downloaded by a background task after the
-           sync.
-        """
-        super(SyncBackgroundTestCase, cls).setUpClass()
-        if cls.cfg.version < Version('2.8'):
-            raise unittest2.SkipTest('This test requires Pulp 2.8 or greater.')
-
-        # Required to ensure content is actually downloaded.
-        utils.reset_squid(cls.cfg)
-        utils.reset_pulp(cls.cfg)
-
-        client = api.Client(cls.cfg, api.json_handler)
-
-        # Create the repository
-        body = gen_repo()
-        body['importer_config'] = {
-            'feed': RPM_FEED_URL,
-            'download_policy': 'background',
-        }
-        distributor = gen_distributor()
-        distributor['auto_publish'] = True
-        distributor['distributor_config']['relative_url'] = body['id']
-        body['distributors'] = [distributor]
-
-        repo = client.post(REPOSITORY_PATH, body)
-        cls.resources.add(repo['_href'])
-
-        # Sync and read the repository
-        sync_path = urljoin(repo['_href'], 'actions/sync/')
-        sync = client.post(sync_path, {'override_config': {}})
-        cls.repo = client.get(repo['_href'], params={'details': True})
-        cls.tasks = tuple(api.poll_spawned_tasks(cls.cfg, sync))
-
-        # Download an RPM to ensure it's satisfied locally.
-        client.response_handler = api.safe_handler
-        path = urljoin('/pulp/repos/', repo['id'] + '/')
-        path = urljoin(path, RPM)
-        cls.rpm = client.get(path)
-
-    def test_repo_local_units(self):
-        """Assert that all content is downloaded for the repository."""
-        self.assertEqual(
-            self.repo['locally_stored_units'],
-            sum(self.repo['content_unit_counts'].values()),
-        )
-
-    def test_request_history(self):
-        """Assert that the request was serviced directly by Pulp.
-
-        If Pulp did not have the content available locally, it would redirect
-        the client to the streamer and the rpm request would contain a history
-        entry for that redirect.
-        """
-        self.assertEqual(0, len(self.rpm.history))
-
-    def test_rpm_checksum(self):
-        """Assert the checksum of the downloaded RPM matches the metadata."""
-        checksum = hashlib.sha256(self.rpm.content).hexdigest()
-        self.assertEqual(RPM_SHA256_CHECKSUM, checksum)
-
-    def test_spawned_download_task(self):
-        """Assert that a download task was spawned as a result of the sync."""
-        expected_tags = {
-            'pulp:repository:' + self.repo['id'],
-            'pulp:action:download',
-        }
-
-        tasks = [t for t in self.tasks if set(t['tags']) == expected_tags]
-        self.assertEqual(1, len(tasks))
-        self.assertEqual('finished', tasks[0]['state'])
 
 
 class RedundantPublishSameMetadataTestCase(utils.BaseAPITestCase):
